+++ src/utils.rs
@@ -37,6 +37,7 @@ pub fn print_config_args(args: &Args) {
     let maketag_string: String = "M".to_owned();
     let ignore_gitcheck_string: String = "X".to_owned();
     let anvil_version_string: String = "a".to_owned();
+    let anvil_kern_string: String = "k".to_owned();
     match args.amboso_dir {
         Some(ref x) => {
             debug!("Passed amboso_dir: {{{}}}", x.display());
@@ -78,6 +79,16 @@ pub fn print_config_args(args: &Args) {
             config_string.push_str(&anvil_version_string);
         }
         None => {}
+    }
+    match args.anvil_kern {
+        Some(ref x) => {
+            debug!("Passed anvil_kern: {{{}}}", x);
+            config_string.push_str(&anvil_kern_string);
+        }
+        None => {
+            trace!("No anvil_kern value");
+            panic!("Missing anvil_kern value");
+        }
     }
     if args.ignore_gitcheck {
         debug!("Ignore git check is on.");